+++ clients/desktop/src/navigation/views.tsx
@@ -17,7 +17,6 @@ import { DepositPage } from '../vault/deposit/DepositPage'
 import EditVaultPage from '../vault/edit/EditVaultPage'
 import { VaultBackupPage } from '../vault/edit/vaultBackupSettings/VaultBackupPage'
 import { ImportVaultFromFilePage } from '../vault/import/components/ImportVaultFromFilePage'
-import { ImportVaultPage } from '../vault/import/components/ImportVaultPage'
 import { JoinKeygenPage } from '../vault/keygen/join/JoinKeygenPage'
 import { SignCustomMessagePage } from '../vault/keysign/customMessage/SignCustomMessagePage'
 import { JoinKeysignPage } from '../vault/keysign/join/JoinKeysignPage'
@@ -38,7 +37,6 @@ const appCustomViews: Views<Exclude<AppViewId, SharedViewId>> = {
   setupFastVault: SetupFastVaultPage,
   setupSecureVault: SetupSecureVaultPage,
   setupVault: SetupVaultPageController,
-  importVault: ImportVaultPage,
   keysign: StartKeysignPage,
   reshareVaultFast: FastReshareVaultPage,
   reshareVaultSecure: SecureReshareVaultPage,
