+++ clients/desktop/src/vault/import/components/ImportVaultFromFilePage.tsx
@@ -1,4 +1,5 @@
-import { ImportVaultFlow } from './ImportVaultFlow'
+import { ImportVaultFlow } from '@core/ui/vault/import/components/ImportVaultFlow'
+
 import { ReadBackupFileStep } from './ReadBackupFileStep'
 
 export const ImportVaultFromFilePage = () => {
