+++ clients/extension/src/navigation/views.tsx
@@ -7,7 +7,6 @@ import { AppViewId } from '@clients/extension/src/navigation/AppView'
 import { ConnectDAppPage } from '@clients/extension/src/pages/connect-dapp'
 import { ConnectedDappsPage } from '@clients/extension/src/pages/connected-dapps'
 import { GetVaultsPage } from '@clients/extension/src/pages/get-vaults'
-import ImportFilePage from '@clients/extension/src/pages/import-file'
 import ImportQRPage from '@clients/extension/src/pages/import-qr'
 import { SettingsPage } from '@clients/extension/src/pages/settings'
 import { SetupVaultPageController } from '@clients/extension/src/pages/setup-vault/SetupVaultPageController'
@@ -26,7 +25,6 @@ const appCustomViews: Views<Exclude<AppViewId, SharedViewId>> = {
   setupFastVault: SetupFastVaultPage,
   setupSecureVault: SetupSecureVaultPage,
   setupVault: SetupVaultPageController,
-  importVault: ImportFilePage,
   keysign: () => <>TODO: Implement keysign page</>,
   reshareVaultFast: ReshareFastVault,
   reshareVaultSecure: ReshareSecureVault,
